pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "linafish"
version = "0.4.0"
description = "Your mind, versioned. The grammar eats your writing and tells you how you think."
requires-python = ">=3.8"
dependencies = []
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Scott Dill", email = "scott@garytcolorado.com"}]
keywords = ["ai", "agent", "codebook", "compression", "mcp", "openclaw", "personalization", "rcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
Homepage = "https://github.com/sdill1973a/ProjectLINA"
Documentation = "https://github.com/sdill1973a/ProjectLINA"
Repository = "https://github.com/sdill1973a/ProjectLINA"

[project.optional-dependencies]
fast = ["numpy>=1.21"]
pdf = ["PyMuPDF"]
docx = ["python-docx"]
gemini = ["google-genai"]
http = ["requests"]
all = ["numpy>=1.21", "PyMuPDF", "python-docx", "google-genai", "requests"]

[project.scripts]
linafish = "linafish.__main__:main"

[tool.setuptools.packages.find]
include = ["linafish*"]

[tool.setuptools.package-data]
linafish = ["data/*.md"]
